move duplicate implementation to central location
