+++ src/gui/input/AxisConfigWidget.cc
@@ -31,9 +31,8 @@
 #include "QSettingsCached.h"
 #include "input/InputDriverManager.h"
 #include "SettingsWriter.h"
-#include "WheelIgnorer.h"
+#include "../IgnoreWheelWhenNotFocused.h"
 #include "InitConfigurator.h"
-
 
 AxisConfigWidget::AxisConfigWidget(QWidget *parent) : QWidget(parent)
 {
@@ -135,17 +134,17 @@ void AxisConfigWidget::init() {
 
   auto comboBoxes = this->findChildren<QComboBox *>();
   if (comboBoxes.size() > 0) { // only allocate if there are comboboxes to use the function
-    WheelIgnorer *wheelIgnorer = new WheelIgnorer(this);
+     auto *ignoreWheelWhenNotFocused = new IgnoreWheelWhenNotFocused(this);
     for (auto comboBox : comboBoxes) {
-      comboBox->installEventFilter(wheelIgnorer); // this takes ownership of the wheelIgnorer object
+      comboBox->installEventFilter(ignoreWheelWhenNotFocused); // this takes ownership of the wheelIgnorer object
       comboBox->setFocusPolicy(Qt::StrongFocus);
     }
   }
   auto spinBoxes = this->findChildren<QDoubleSpinBox *>();
   if (spinBoxes.size() > 0) { // only allocate if there are spinboxes to use the function
-    auto *wheelIgnorer = new WheelIgnorer(this);
+    auto *ignoreWheelWhenNotFocused = new IgnoreWheelWhenNotFocused(this);
     for (auto spinBox : spinBoxes) {
-      spinBox->installEventFilter(wheelIgnorer); // this takes ownership of the wheelIgnorer object
+      spinBox->installEventFilter(ignoreWheelWhenNotFocused); // this takes ownership of the wheelIgnorer object
       spinBox->setFocusPolicy(Qt::StrongFocus);
     }
   }
