+++ src/gui/parameter/ParameterVector.cc
@@ -1,5 +1,5 @@
 #include "ParameterVector.h"
-#include "IgnoreWheelWhenNotFocused.h"
+#include "../IgnoreWheelWhenNotFocused.h"
 
 ParameterVector::ParameterVector(QWidget *parent, VectorParameter *parameter, DescriptionStyle descriptionStyle) :
   ParameterVirtualWidget(parent, parameter),
